removed outdated scripts + restructure
