+++ config/go2.json
@@ -18,21 +18,89 @@
     "tau_0": 0.1,
     "tau_1": 0.0,
     "imu_rate": 500.0,
-    "R_base_to_gyro": [1.0, 0.0, 0.0, 0.0, -1.0, 0.0, 0.0, 0.0, -1.0],
-    "R_base_to_acc": [1.0, 0.0, 0.0, 0.0, -1.0, 0.0, 0.0, 0.0, -1.0],
+    "R_base_to_gyro": [
+        1.0,
+        0.0,
+        0.0,
+        0.0,
+        -1.0,
+        0.0,
+        0.0,
+        0.0,
+        -1.0
+    ],
+    "R_base_to_acc": [
+        1.0,
+        0.0,
+        0.0,
+        0.0,
+        -1.0,
+        0.0,
+        0.0,
+        0.0,
+        -1.0
+    ],
     "calibrate_initial_imu_bias": true,
     "max_imu_calibration_cycles": 1000,
-    "bias_gyro": [0.0, 0.0, 0.0],
-    "bias_acc": [0.0, 0.0, 0.0],
+    "bias_gyro": [
+        0.0,
+        0.0,
+        0.0
+    ],
+    "bias_acc": [
+        0.0,
+        0.0,
+        0.0
+    ],
     "gyro_cutoff_frequency": 5.0,
     "force_torque_rate": 500.0,
     "R_foot_to_force": {
-        "0": [1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0],
-        "1": [1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0],
-        "2": [1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0],
-        "3": [1.0, 0.0, 0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 1.0]
+        "0": [
+            1.0,
+            0.0,
+            0.0,
+            0.0,
+            1.0,
+            0.0,
+            0.0,
+            0.0,
+            1.0
+        ],
+        "1": [
+            1.0,
+            0.0,
+            0.0,
+            0.0,
+            1.0,
+            0.0,
+            0.0,
+            0.0,
+            1.0
+        ],
+        "2": [
+            1.0,
+            0.0,
+            0.0,
+            0.0,
+            1.0,
+            0.0,
+            0.0,
+            0.0,
+            1.0
+        ],
+        "3": [
+            1.0,
+            0.0,
+            0.0,
+            0.0,
+            1.0,
+            0.0,
+            0.0,
+            0.0,
+            1.0
+        ]
     },
-    "R_foot_to_torque": null,   
+    "R_foot_to_torque": null,
     "attitude_estimator_proportional_gain": 0.1,
     "attitude_estimator_integral_gain": 0.0,
     "estimate_contact_status": true,
@@ -41,36 +109,121 @@
     "median_window": 13,
     "outlier_detection": true,
     "convergence_cycles": 100,
-    "imu_angular_velocity_covariance": [5e-5, 5e-5, 5e-5],
-    "imu_angular_velocity_bias_covariance": [1e-8, 1e-8, 1e-8],
-    "imu_linear_acceleration_covariance": [5e-3, 5e-3, 5e-3],
-    "imu_linear_acceleration_bias_covariance": [1e-8, 1e-8, 1e-8],
-    
+    "imu_angular_velocity_covariance": [
+        5e-5,
+        5e-5,
+        5e-5
+    ],
+    "imu_angular_velocity_bias_covariance": [
+        1e-8,
+        1e-8,
+        1e-8
+    ],
+    "imu_linear_acceleration_covariance": [
+        5e-3,
+        5e-3,
+        5e-3
+    ],
+    "imu_linear_acceleration_bias_covariance": [
+        1e-8,
+        1e-8,
+        1e-8
+    ],
     "use_contacts_in_base_estimation": true,
-    "base_linear_velocity_covariance": [5e-3, 5e-3, 5e-3],
-    "base_orientation_covariance": [1e-3, 1e-3, 1e-3],
-
-    "contact_position_covariance": [1e-6, 1e-6, 1e-6],
+    "base_linear_velocity_covariance": [
+        5e-3,
+        5e-3,
+        5e-3
+    ],
+    "base_orientation_covariance": [
+        1e-3,
+        1e-3,
+        1e-3
+    ],
+    "contact_position_covariance": [
+        1e-6,
+        1e-6,
+        1e-6
+    ],
     "contact_orientation_covariance": null,
-    "contact_position_slip_covariance": [1e-6, 1e-6, 1e-6],
+    "contact_position_slip_covariance": [
+        1e-6,
+        1e-6,
+        1e-6
+    ],
     "contact_orientation_slip_covariance": null,
-    "com_position_process_covariance": [1e-6, 1e-6, 1e-6],
-    "com_linear_velocity_process_covariance": [1e-4, 1e-4, 1e-4],
-    "external_forces_process_covariance": [1e-1, 1e-1, 1e-1],
-    "com_position_covariance": [1e-6, 1e-6, 1e-6],
-    "com_linear_acceleration_covariance": [1e-2, 1e-2, 1e-2],
-    "initial_base_position_covariance": [1e-4, 1e-4, 1e-4],
-    "initial_base_orientation_covariance": [1e-6, 1e-6, 1e-6],
-    "initial_base_linear_velocity_covariance": [1e-3, 1e-3, 1e-3],
-    "initial_contact_position_covariance": [1e-6, 1e-6, 1e-6],
+    "com_position_process_covariance": [
+        1e-6,
+        1e-6,
+        1e-6
+    ],
+    "com_linear_velocity_process_covariance": [
+        1e-4,
+        1e-4,
+        1e-4
+    ],
+    "external_forces_process_covariance": [
+        1e-1,
+        1e-1,
+        1e-1
+    ],
+    "com_position_covariance": [
+        1e-6,
+        1e-6,
+        1e-6
+    ],
+    "com_linear_acceleration_covariance": [
+        1e-2,
+        1e-2,
+        1e-2
+    ],
+    "initial_base_position_covariance": [
+        1e-4,
+        1e-4,
+        1e-4
+    ],
+    "initial_base_orientation_covariance": [
+        1e-6,
+        1e-6,
+        1e-6
+    ],
+    "initial_base_linear_velocity_covariance": [
+        1e-3,
+        1e-3,
+        1e-3
+    ],
+    "initial_contact_position_covariance": [
+        1e-6,
+        1e-6,
+        1e-6
+    ],
     "initial_contact_orientation_covariance": null,
-    "initial_imu_linear_acceleration_bias_covariance": [1e-4, 1e-4, 1e-4],
-    "initial_imu_angular_velocity_bias_covariance": [1e-4, 1e-4, 1e-4],
-    "initial_com_position_covariance": [1e-6, 1e-6, 1e-6],
-    "initial_com_linear_velocity_covariance": [1e-3, 1e-3, 1e-3],
-    "initial_external_forces_covariance": [1.0, 1.0, 1.0],
+    "initial_imu_linear_acceleration_bias_covariance": [
+        1e-4,
+        1e-4,
+        1e-4
+    ],
+    "initial_imu_angular_velocity_bias_covariance": [
+        1e-4,
+        1e-4,
+        1e-4
+    ],
+    "initial_com_position_covariance": [
+        1e-6,
+        1e-6,
+        1e-6
+    ],
+    "initial_com_linear_velocity_covariance": [
+        1e-3,
+        1e-3,
+        1e-3
+    ],
+    "initial_external_forces_covariance": [
+        1.0,
+        1.0,
+        1.0
+    ],
     "T_base_to_odom": null,
-
-    "enable_terrain_estimation": false,
+    "enable_terrain_estimation": true,
     "minimum_terrain_height_variance": 1e-3
 }
