+++ evaluation/mujoco_test/test_config.json
@@ -1,0 +1,11 @@
+{
+  "Experiment": {
+    "type": "rough"
+  },
+  "Paths": {
+    "base_path": "/evaluation/mujoco_test/mujoco_data",
+    "data_file": "{base_path}/{type}/go2_data.h5",
+    "prediction_file": "{base_path}/{type}/serow_predictions.h5",
+    "elevation_map_file": "{base_path}/{type}/est_elevation_map.bin"
+  }
+}
