--- ros2/serow_ros2/package.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>serow_ros2</name>
-  <version>1.0.0</version>
-  <description>State Estimation Robot Walking (SEROW) ros2 driver</description>
-  <maintainer email="stpiperakis@gmail.com">Stylianos Piperakis</maintainer>
-  <maintainer email="maravgakis@ics.forth.gr">Micheal Maravgakis</maintainer>
-
-  <license>GPL-3.0-only</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <depend>rclcpp</depend>
-  <depend>geometry_msgs</depend>
-  <depend>sensor_msgs</depend>
-  <depend>yaml-cpp</depend>
-
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>ament_lint_common</test_depend>
-  
-  <exec_depend>ros2launch</exec_depend>
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
