--- test/scripts/plot_grf_config.yaml
@@ -1,3 +1,0 @@
-data_filename: "../data/go1.csv"  # Dataset containing GRF measurements
-
-force_column_indxs: [1,2,3,4]      # Columns in the CSV that contain GRFs.
